refactor: improve cancellation sheet creation
